refactor: rename Nav in MenuItems
