+++ frontend/svelte/src/lib/components/common/Menu.svelte
@@ -2,7 +2,7 @@
   import { i18n } from "../../stores/i18n";
   import Menu from "../ui/Menu.svelte";
   import GetICPs from "../ic/GetICPs.svelte";
-  import Nav from "./Nav.svelte";
+  import MenuItems from "./MenuItems.svelte";
   import { IS_TESTNET } from "../../constants/environment.constants";
   import IconMenu from "../../icons/IconMenu.svelte";
   import IconClose from "../../icons/IconClose.svelte";
@@ -27,7 +27,7 @@
     class="close icon-only"><IconClose /></button
   >
 
-  <Nav />
+  <MenuItems />
 
   {#if IS_TESTNET}
     <GetICPs />
