+++ frontend/svelte/src/lib/components/common/MenuItems.svelte
